+++ src/app/(dashboard)/dashboard/settings/components/modalityBridge/ModalityBridgeVideoTab.tsx
@@ -26,13 +26,23 @@ interface VideoState {
   modalityBridgeVideoTimeout: number;
 }
 
-interface RuntimeStatus {
-  available: boolean;
-  ffmpegVersion: string | null;
-  ffprobeVersion: string | null;
-  reason?: string;
-  restricted?: boolean;
-}
+// Explicit UI states for the FFmpeg/ffprobe runtime probe (#11657):
+//  - "unknown"     — not yet checked, or the check could not be completed
+//                    (in-flight fetch, network error, non-2xx response other
+//                    than the loopback classification below). MUST NOT be
+//                    presented as "unavailable" — that would tell an operator
+//                    to install FFmpeg when the real cause may be transient.
+//  - "restricted"  — this dashboard host is not loopback, so the probe is
+//                    intentionally skipped (client-side classification, no
+//                    request sent).
+//  - "unavailable" — the probe ran and confirmed FFmpeg/ffprobe are missing
+//                    or failed to start.
+//  - "available"   — the probe ran and confirmed a working runtime.
+type RuntimeStatus =
+  | { state: "unknown" }
+  | { state: "restricted" }
+  | { state: "unavailable"; reason?: string }
+  | { state: "available"; ffmpegVersion: string | null; ffprobeVersion: string | null };
 
 interface ModalityBridgeVideoTabProps {
   runtimeHostname?: string;
@@ -55,21 +65,22 @@ function fromApi(value: unknown): VideoState {
   };
 }
 
-function parseRuntimeStatus(value: unknown): RuntimeStatus | null {
+function parseRuntimeStatus(value: unknown): RuntimeStatus {
   const record = asRecord(value);
-  if (record.restricted === true) {
+  if (record.restricted === true) return { state: "restricted" };
+  // A missing/malformed `available` field means the probe could not be
+  // completed (in-flight, network failure, or an unexpected response shape)
+  // — that is "unknown", never "unavailable".
+  if (typeof record.available !== "boolean") return { state: "unknown" };
+  if (record.available) {
     return {
-      available: false,
-      ffmpegVersion: null,
-      ffprobeVersion: null,
-      restricted: true,
+      state: "available",
+      ffmpegVersion: typeof record.ffmpegVersion === "string" ? record.ffmpegVersion : null,
+      ffprobeVersion: typeof record.ffprobeVersion === "string" ? record.ffprobeVersion : null,
     };
   }
-  if (typeof record.available !== "boolean") return null;
   return {
-    available: record.available,
-    ffmpegVersion: typeof record.ffmpegVersion === "string" ? record.ffmpegVersion : null,
-    ffprobeVersion: typeof record.ffprobeVersion === "string" ? record.ffprobeVersion : null,
+    state: "unavailable",
     reason: typeof record.reason === "string" ? record.reason : undefined,
   };
 }
@@ -95,7 +106,7 @@ export default function ModalityBridgeVideoTab({
   const t = useTranslations("settings");
   const tRoot = useTranslations();
   const [settings, setSettings] = useState<VideoState | null>(null);
-  const [runtime, setRuntime] = useState<RuntimeStatus | null>(null);
+  const [runtime, setRuntime] = useState<RuntimeStatus>({ state: "unknown" });
   const [errorState, setErrorState] = useState<"load" | "save" | null>(null);
   const persistedSettings = useRef<VideoState | null>(null);
   const isVisionModel = useCallback((model: ApiModel) => model.supportsVision === true, []);
@@ -191,30 +202,34 @@ export default function ModalityBridgeVideoTab({
         ) : null}
         <div
           className={`rounded-control border p-3 text-sm ${
-            runtime?.available
+            runtime.state === "available"
               ? "border-success/30 bg-success/5 text-success"
               : "border-warning/30 bg-warning/5 text-text-muted"
           }`}
           aria-live="polite"
         >
-          {runtime?.available ? (
+          {runtime.state === "available" ? (
             <>
               <strong>{t("modalityBridgeVideoRuntimeReady")}</strong>
               <div className="mt-1 text-xs">
                 FFmpeg {runtime.ffmpegVersion} · ffprobe {runtime.ffprobeVersion}
               </div>
             </>
-          ) : runtime?.restricted ? (
+          ) : runtime.state === "restricted" ? (
             <>
               <strong>{t("authz.badge.strict")}</strong>
               <div className="mt-1 text-xs">{tRoot("endpoint.badgeLoopbackTooltip")}</div>
             </>
-          ) : (
+          ) : runtime.state === "unavailable" ? (
             <>
               <strong>{t("modalityBridgeVideoRuntimeUnavailable")}</strong>
               <div className="mt-1 text-xs">
-                {runtime?.reason || t("modalityBridgeVideoRuntimeInstall")}
+                {runtime.reason || t("modalityBridgeVideoRuntimeInstall")}
               </div>
+            </>
+          ) : (
+            <>
+              <strong>{t("modalityBridgeVideoRuntimeChecking")}</strong>
             </>
           )}
         </div>
