+++ src/i18n/messages/ar.json
@@ -8186,6 +8186,7 @@
     "modalityBridgeVideoTitle": "جسر الفيديو",
     "modalityBridgeVideoDesc": "خذ عينات من إطارات الفيديو وصفها باستخدام نموذج رؤية، ثم تابع باستخدام نموذج النص المحدد.",
     "modalityBridgeVideoRuntimeReady": "بيئة التشغيل جاهزة",
+    "modalityBridgeVideoRuntimeChecking": "جارٍ فحص وقت التشغيل…",
     "modalityBridgeVideoRuntimeUnavailable": "بيئة التشغيل غير متاحة",
     "modalityBridgeVideoRuntimeInstall": "ثبّت FFmpeg وffprobe في PATH الخاص بالخادم لتمكين استخراج الفيديو.",
     "modalityBridgeVideoEnabled": "تفعيل جسر الفيديو",
