+++ src/i18n/messages/in.json
@@ -8186,6 +8186,7 @@
     "modalityBridgeVideoTitle": "Jembatan Video",
     "modalityBridgeVideoDesc": "Ambil sampel bingkai video, deskripsikan dengan model visi, lalu lanjutkan dengan model teks yang dipilih.",
     "modalityBridgeVideoRuntimeReady": "Runtime siap",
+    "modalityBridgeVideoRuntimeChecking": "Memeriksa runtime…",
     "modalityBridgeVideoRuntimeUnavailable": "Runtime tidak tersedia",
     "modalityBridgeVideoRuntimeInstall": "Pasang FFmpeg dan ffprobe di PATH server untuk mengaktifkan ekstraksi video.",
     "modalityBridgeVideoEnabled": "Aktifkan Jembatan Video",
