+++ src/i18n/messages/ko.json
@@ -8186,6 +8186,7 @@
     "modalityBridgeVideoTitle": "비디오 브리지",
     "modalityBridgeVideoDesc": "비디오 프레임을 샘플링하고 비전 모델로 설명한 다음 선택한 텍스트 모델로 계속 처리합니다.",
     "modalityBridgeVideoRuntimeReady": "런타임 준비 완료",
+    "modalityBridgeVideoRuntimeChecking": "런타임 확인 중…",
     "modalityBridgeVideoRuntimeUnavailable": "런타임을 사용할 수 없음",
     "modalityBridgeVideoRuntimeInstall": "비디오 추출을 사용하려면 서버의 PATH에 FFmpeg와 ffprobe를 설치하세요.",
     "modalityBridgeVideoEnabled": "비디오 브리지 사용",
