+++ src/i18n/messages/ro.json
@@ -8186,6 +8186,7 @@
     "modalityBridgeVideoTitle": "Punte video",
     "modalityBridgeVideoDesc": "Eșantionați cadre video, descrieți-le cu un model vizual și continuați cu modelul text ales.",
     "modalityBridgeVideoRuntimeReady": "Mediul de execuție este pregătit",
+    "modalityBridgeVideoRuntimeChecking": "Se verifică mediul de execuție…",
     "modalityBridgeVideoRuntimeUnavailable": "Mediul de execuție nu este disponibil",
     "modalityBridgeVideoRuntimeInstall": "Instalați FFmpeg și ffprobe în PATH-ul serverului pentru a activa extragerea video.",
     "modalityBridgeVideoEnabled": "Activați Punte video",
