+++ src/i18n/messages/ta.json
@@ -8186,6 +8186,7 @@
     "modalityBridgeVideoTitle": "வீடியோ பாலம்",
     "modalityBridgeVideoDesc": "வீடியோ சட்டகங்களை மாதிரியாக எடுத்து, பார்வை மாதிரியால் அவற்றை விவரித்து, தேர்ந்தெடுத்த உரை மாதிரியுடன் தொடரவும்.",
     "modalityBridgeVideoRuntimeReady": "இயக்கச்சூழல் தயாராக உள்ளது",
+    "modalityBridgeVideoRuntimeChecking": "இயக்க நேரத்தைச் சரிபார்க்கிறது…",
     "modalityBridgeVideoRuntimeUnavailable": "இயக்கச்சூழல் கிடைக்கவில்லை",
     "modalityBridgeVideoRuntimeInstall": "வீடியோ பிரித்தெடுப்பைச் செயல்படுத்த சேவையகத்தின் PATH இல் FFmpeg மற்றும் ffprobe ஐ நிறுவவும்.",
     "modalityBridgeVideoEnabled": "வீடியோ பாலத்தைச் செயல்படுத்து",
